+++ protected/humhub/modules/post/widgets/views/form.php
@@ -9,12 +9,14 @@ use humhub\modules\ui\form\widgets\ActiveForm;
 /* @var ActiveForm $form */
 /* @var Post $post */
 /* @var string $submitUrl */
+/* @var $fileList array */
+/* @var $isModal bool */
 ?>
 
 <?= $form->field($post, 'message')->widget(RichTextField::class, [
     'id' => 'contentForm_message',
     'form' => $form,
-    'layout' => RichTextField::LAYOUT_INLINE,
+    'layout' => $isModal ? RichTextField::LAYOUT_BLOCK : RichTextField::LAYOUT_INLINE,
     'pluginOptions' => ['maxHeight' => '300px'],
     'placeholder' => Yii::t("PostModule.base", "What's on your mind?"),
     'name' => 'message',
@@ -26,4 +28,6 @@ use humhub\modules\ui\form\widgets\ActiveForm;
 <?= WallCreateContentFormFooter::widget([
     'contentContainer' => $post->content->container,
     'submitUrl' => $submitUrl,
+    'fileList' => $fileList,
+    'isModal' => $isModal,
 ]) ?>